Moved PosProfitabilityScore into component
